using 'unit256' consistently instead of mixed with 'uint'

+++ server/contracts/Core/RootChain.sol
@@ -18,16 +18,16 @@ import "./SparseMerkleTree.sol";
 
 contract RootChainEvents {
     event Deposit(uint64 indexed slot, uint256 depositBlockNumber, uint64 denomination, address indexed from, bytes32 hash);
-    event SubmittedBlock(uint blockNumber, bytes32 root, uint timestamp);
+    event SubmittedBlock(uint256 blockNumber, bytes32 root, uint256 timestamp);
 
-    event StartedExit(uint64 indexed slot, address indexed owner, uint createdAt);
+    event StartedExit(uint64 indexed slot, address indexed owner, uint256 createdAt);
     event ChallengedExit(uint64 indexed slot);
     event RespondedExitChallenge(uint64 indexed slot);
     event FinalizedExit(uint64  indexed slot, address owner);
 
-    event FreedBond(address indexed from, uint amount);
-    event SlashedBond(address indexed from, address indexed to, uint amount);
-    event WithdrewBonds(address indexed from, uint amount);
+    event FreedBond(address indexed from, uint256 amount);
+    event SlashedBond(address indexed from, address indexed to, uint256 amount);
+    event WithdrewBonds(address indexed from, uint256 amount);
 }
 
 
@@ -37,7 +37,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     using Transaction for bytes;
     using ECVerify for bytes32;
 
-    uint constant BOND_AMOUNT = 0.1 ether;
+    uint256 constant BOND_AMOUNT = 0.1 ether;
 
     address public authority;
 
@@ -71,8 +71,8 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     }
 
     struct Balance {
-        uint bonded;
-        uint withdrawable;
+        uint256 bonded;
+        uint256 withdrawable;
     }
     mapping (address => Balance ) public balances;
 
@@ -117,11 +117,11 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         uint256 createdAt;
     }
 
-    uint public depositCount;
-    mapping(uint => childBlock) public childChain;
+    uint256 public depositCount;
+    mapping(uint256 => childBlock) public childChain;
     CryptoCards cryptoCards;
 
-    constructor () public{
+    constructor () public {
         authority = msg.sender;
         childBlockInterval = 1000;
         currentChildBlock = childBlockInterval;
@@ -187,7 +187,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         uint64 slot, bytes prevTxBytes,
         bytes exitingTxBytes, bytes prevTxInclusionProof,
         bytes exitingTxInclusionProof, bytes sig,
-        uint prevTxIncBlock, uint exitingTxIncBlock)
+        uint256 prevTxIncBlock, uint256 exitingTxIncBlock)
         isState(slot, State.DEPOSITED)
         payable isBonded
         external
@@ -245,8 +245,8 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     }
 
     function finalizeExits() external {
-        uint exitSlotsLength = exitSlots.length;
-        for (uint i = 0; i < exitSlotsLength; i++) {
+        uint256 exitSlotsLength = exitSlots.length;
+        for (uint256 i = 0; i < exitSlotsLength; i++) {
             finalizeExit(exitSlots[i]);
         }
     }
@@ -266,7 +266,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         bytes prevTxBytes, bytes exitingTxBytes,
         bytes prevTxInclusionProof, bytes exitingTxInclusionProof,
         bytes sig,
-        uint prevTxIncBlock, uint exitingTxIncBlock)
+        uint256 prevTxIncBlock, uint256 exitingTxIncBlock)
         external
         payable isBonded
         isState(slot, State.EXITING)
@@ -284,7 +284,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     }
 
     // If `challengeBefore` was successfully challenged, then set state to RESPONDED and allow the coin to be exited. No need to actually attach a bond when responding to a challenge
-    function respondChallengeBefore(uint64 slot, uint challengingBlockNumber, bytes challengingTransaction, bytes proof)
+    function respondChallengeBefore(uint64 slot, uint256 challengingBlockNumber, bytes challengingTransaction, bytes proof)
         external
         isState(slot, State.CHALLENGED)
     {
@@ -296,7 +296,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     }
 
 
-    function challengeBetween(uint64 slot, uint challengingBlockNumber, bytes challengingTransaction, bytes proof)
+    function challengeBetween(uint64 slot, uint256 challengingBlockNumber, bytes challengingTransaction, bytes proof)
         external isState(slot, State.EXITING) cleanupExit(slot)
     {
         // Must challenge with a tx in between
@@ -310,7 +310,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         coins[slot].state = State.DEPOSITED;
     }
 
-    function challengeAfter(uint64 slot, uint challengingBlockNumber, bytes challengingTransaction, bytes proof)
+    function challengeAfter(uint64 slot, uint256 challengingBlockNumber, bytes challengingTransaction, bytes proof)
         external
         isState(slot, State.EXITING)
         cleanupExit(slot)
@@ -334,7 +334,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
 
     function withdrawBonds() external {
         // Can only withdraw bond if the msg.sender
-        uint amount = balances[msg.sender].withdrawable;
+        uint256 amount = balances[msg.sender].withdrawable;
         balances[msg.sender].withdrawable = 0; // no reentrancy!
 
         msg.sender.transfer(amount);
@@ -380,7 +380,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
     function checkDepositBlockInclusion(
         bytes txBytes,
         bytes signature,
-        uint txIncBlock,
+        uint256 txIncBlock,
         bool checkSender
     )
          private
@@ -405,7 +405,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         bytes prevTxBytes, bytes exitingTxBytes,
         bytes prevTxInclusionProof, bytes exitingTxInclusionProof,
         bytes sig,
-        uint prevTxIncBlock, uint exitingTxIncBlock, bool checkSender)
+        uint256 prevTxIncBlock, uint256 exitingTxIncBlock, bool checkSender)
         private
         view
         returns (bool)
@@ -424,7 +424,7 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         return true;
     }
 
-    function checkTxIncluded(bytes txBytes, uint blockNumber, bytes proof) private view {
+    function checkTxIncluded(bytes txBytes, uint256 blockNumber, bytes proof) private view {
         Transaction.TX memory txData = txBytes.getTx();
         bytes32 txHash;
         bytes32 root = childChain[blockNumber].root;
@@ -465,9 +465,9 @@ contract RootChain is ERC721Receiver, SparseMerkleTree, RootChainEvents {
         cryptoCards = _cryptoCards;
     }
 
-    function getIndex(uint64 slot) private view returns (uint) {
-        uint len = exitSlots.length;
-        for (uint i = 0; i < len; i++) {
+    function getIndex(uint64 slot) private view returns (uint256) {
+        uint256 len = exitSlots.length;
+        for (uint256 i = 0; i < len; i++) {
             if (exitSlots[i] == slot)
                 return i;
         }
